TIL how to write operating-system specific aliasses
I've ran into some issues regarding aliasses that either work
on Linux or macOS but not on both for quite some time.
Since i use my zshell config on most system i used to downgrade
to the "least common nominator" - not defining the alias at all.

But with this, things change.